requirements.txt
google-generativeai==0.5.4
openai==0.18.1
openpyxl==3.1.2